7 products.php http://pm.5psolutions.net/public/index.php?path_info=projects/training-razvan-mocanu/tasks/7 added delete restriction if products are in order

+++ products.php
@@ -2,22 +2,32 @@
 
 require_once 'common.php';
 $selectProducts = 'SELECT * FROM products';
-$pdoConnection = getDatabaseConnection();
 $products = getProductsArray($pdoConnection, $selectProducts);
 
 checkLogin();
 
 if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-    echo $_POST['delete'];
     if (isset($_POST['delete'])) {
-        $deleteProduct = $pdoConnection->prepare('DELETE FROM products WHERE id=? LIMIT 1');
-        if ($deleteProduct->execute([strip_tags($_POST['delete'])]) === TRUE) {
-            $success['delete'] = 'Record deleted successfully!';
-        } else {
-            $failure['delete'] = 'Failed deleting record!';
+        $selectProduct = $pdoConnection->prepare('SELECT OP.id_order
+                                                                   FROM products P
+                                                                   INNER JOIN order_product OP ON OP.id_product = P.id 
+                                                                   INNER JOIN orders O ON O.id = OP.id_order WHERE P.id = ? GROUP BY OP.id_product');
+        $selectProduct->execute([strip_tags($_POST['delete'])]);
+
+        if(empty($selectProduct->fetch())){
+            $deleteProduct = $pdoConnection->prepare('DELETE FROM products WHERE id=? LIMIT 1');
+            if ($deleteProduct->execute([strip_tags($_POST['delete'])]) === TRUE) {
+                $success['delete'] = 'Record deleted successfully!';
+            } else {
+                $failure['delete'][$_POST['delete']] = 'Failed deleting record from table!';
+            }
+            header('Location: products.php');
+            die();
         }
-        header('Location: products.php');
-        die();
+        else {
+            $failure['delete'][$_POST['delete']] = 'Failed deleting record, there are orders containing it!';
+        }
+
     }
 
     if (isset($_POST['logout'])) {
@@ -36,9 +46,14 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
     <link rel="stylesheet" href="stylesheets/index.css">
 </head>
 <body>
-<?= isset($success['delete']) ? translateText($success['delete']) : '' ?>
-<?= isset($failure['delete']) ? translateText($failure['delete']) : '' ?>
+<a href="product.php?addProduct=true"><?= translateText('Add Product') ?> </a>
+<a href="orders.php"><?= translateText('Go to orders') ?></a>
+<form action="products.php" method="post">
+    <input type="hidden" name="logout" value="true">
+    <button type="submit"><?= translateText('Logout') ?></button>
+</form>
 <h1> <?= translateText('Welcome back, ') . $_SESSION['username'] ?> !</h1>
+<div class="success"><?= isset($success['delete']) ? translateText($success['delete']) : '' ?></div>
 <?php foreach ($products as $product): ?>
     <div class="product">
         <img src="images/<?= $product['id'] ?>.png" alt="<?= $product['id'] ?>-image"
@@ -59,10 +74,12 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
             </form>
         </span>
     </div>
+    <div class="error"><?= isset($failure['delete'][$product['id']]) ? translateText($failure['delete'][$product['id']]) : '' ?></div>
+
     <br>
 <?php endforeach ?>
 <a href="product.php?addProduct=true"><?= translateText('Add Product') ?> </a>
-
+<a href="orders.php"><?= translateText('Go to orders') ?></a>
 <form action="products.php" method="post">
     <input type="hidden" name="logout" value="true">
     <button type="submit"><?= translateText('Logout') ?></button>
